flaskapi/requirements.txt
Flask==3.0.0
opencv-python>=4.8.0
ultralytics>=8.4.0
insightface>=0.7.3
onnxruntime>=1.16.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests==2.31.0
python-dotenv==1.0.0
